ColorRenderNode🔥
- RenderNodes now only do the bindings, set push constants if used,
  record the draw call and end the rendering when recording command
  buffers.

- RenderNodes also have now `perFrameFunctions` which are up to 3
  functions that can be injected before starting the bindings and ending
  the rendering. These functions needs to start the dynamica rendering
  or it won't work.

- RenderNodes now owns 1 VImage as input and another as output for doing
  multi-node configurations in the future.

- ColorRenderNode: Meant to be the first or last node in the render
  graph. If `present` is true, it will attach the swapchain image as
  color attachment, render to it and transition it to present. If
  `present` is not true, it will initialize a new VImage as output, use
  it as color attachment, render to it and transition it to be used as a
  sampled image in a shader.

- Finaly using things like the importer with this new renderer and
  actually seeing it on the screen. Vertex now uses the template needed
  for RenderNode to actually get bindings an attribs.

+++ src/engine/blender/importer.h
@@ -59,7 +59,7 @@ struct BlenderModel {
   std::string name;
   float fps = 24.0f;
   std::vector<Vertex> vertices;
-  std::vector<std::uint16_t> indices;
+  std::vector<std::uint32_t> indices;
   std::vector<AnimationClip> animations;
 };
 
@@ -109,7 +109,7 @@ struct BlenderTransformModel {
   std::string name;
   float fps = 24.0f;
   std::vector<Vertex> vertices;
-  std::vector<std::uint16_t> indices;
+  std::vector<std::uint32_t> indices;
   std::vector<TransformAnimationClip> animations;
 };
